requirements.txt
Django==1.8c1
argparse==1.2.1
django-cors-headers==1.0.0
djangorestframework==3.1.1
psycopg2==2.6
wsgiref==0.1.2
